refactor(gradle): move subproject to root
