+++ src/main/AndroidManifest.xml
